requirements.txt
confirparser
numpy
pyqt==5